Move spec metadata into metadata package
