+++ core/statix.multilang/src/main/java/mb/statix/multilang/pie/SmlPartialSolveProject.java
@@ -8,7 +8,7 @@ import mb.pie.api.TaskDef;
 import mb.statix.constraints.CUser;
 import mb.statix.multilang.MultiLangAnalysisException;
 import mb.statix.multilang.MultiLangScope;
-import mb.statix.multilang.spec.SpecLoadException;
+import mb.statix.multilang.metadata.spec.SpecLoadException;
 import mb.statix.multilang.utils.SolverUtils;
 import mb.statix.solver.IConstraint;
 import mb.statix.solver.log.IDebugContext;
