+++ core/statix.multilang/src/main/java/mb/statix/multilang/pie/SmlSolveProject.java
@@ -19,7 +19,7 @@ import mb.statix.multilang.metadata.LanguageMetadataManager;
 import mb.statix.multilang.MultiLang;
 import mb.statix.multilang.MultiLangAnalysisException;
 import mb.statix.multilang.MultiLangScope;
-import mb.statix.multilang.spec.SpecLoadException;
+import mb.statix.multilang.metadata.spec.SpecLoadException;
 import mb.statix.solver.IConstraint;
 import mb.statix.solver.IState;
 import mb.statix.solver.log.IDebugContext;
