+++ core/statix.multilang/src/main/java/mb/statix/multilang/utils/MetadataUtils.java
@@ -9,9 +9,9 @@ import mb.resource.hierarchical.match.PathResourceMatcher;
 import mb.resource.hierarchical.match.path.ExtensionsPathMatcher;
 import mb.resource.hierarchical.match.path.NoHiddenPathMatcher;
 import mb.resource.hierarchical.walk.PathResourceWalker;
-import mb.statix.multilang.spec.SpecBuilder;
-import mb.statix.multilang.spec.SpecLoadException;
-import mb.statix.multilang.spec.SpecUtils;
+import mb.statix.multilang.metadata.spec.SpecBuilder;
+import mb.statix.multilang.metadata.spec.SpecLoadException;
+import mb.statix.multilang.metadata.spec.SpecUtils;
 import org.spoofax.interpreter.terms.ITermFactory;
 
 import java.io.IOException;
